Refactor Imports and Project Structure
