+++ brain_brew/build_tasks/crowd_anki/crowd_anki_generate.py
@@ -5,8 +5,8 @@ from brain_brew.build_tasks.crowd_anki.headers_to_crowd_anki import HeadersToCro
 from brain_brew.build_tasks.crowd_anki.media_to_crowd_anki import MediaGroupToCrowdAnki
 from brain_brew.build_tasks.crowd_anki.note_models_to_crowd_anki import NoteModelsToCrowdAnki
 from brain_brew.build_tasks.crowd_anki.notes_to_crowd_anki import NotesToCrowdAnki
-from brain_brew.representation.build_config.build_task import TopLevelBuildTask
-from brain_brew.representation.configuration.representation_base import RepresentationBase
+from brain_brew.configuration.build_config.build_task import TopLevelBuildTask
+from brain_brew.configuration.representation_base import RepresentationBase
 from brain_brew.representation.generic.media_file import MediaFile
 from brain_brew.representation.json.crowd_anki_export import CrowdAnkiExport
 from brain_brew.representation.json.wrappers_for_crowd_anki import CrowdAnkiJsonWrapper
