+++ brain_brew/build_tasks/crowd_anki/media_group_from_crowd_anki.py
@@ -2,9 +2,9 @@ from dataclasses import dataclass
 from typing import Union
 
 from brain_brew.build_tasks.deck_parts.media_group_from_folder import MediaGroupFromFolder
+from brain_brew.configuration.part_holder import PartHolder
 from brain_brew.representation.json.crowd_anki_export import CrowdAnkiExport
-from brain_brew.representation.yaml.media_group_repr import MediaGroup
-from brain_brew.representation.yaml.part_holder import PartHolder
+from brain_brew.representation.yaml.media_group import MediaGroup
 from brain_brew.transformers.media_group_from_location import create_media_group_from_location
