+++ brain_brew/build_tasks/crowd_anki/note_models_from_crowd_anki.py
@@ -1,13 +1,12 @@
 from dataclasses import dataclass, field
 from typing import Optional, Union
 
-from brain_brew.representation.build_config.build_task import BuildPartTask
-
-from brain_brew.representation.configuration.representation_base import RepresentationBase
+from brain_brew.configuration.build_config.build_task import BuildPartTask
+from brain_brew.configuration.part_holder import PartHolder
+from brain_brew.configuration.representation_base import RepresentationBase
 from brain_brew.representation.json.crowd_anki_export import CrowdAnkiExport
 from brain_brew.representation.json.wrappers_for_crowd_anki import CrowdAnkiJsonWrapper
-from brain_brew.representation.yaml.note_model_repr import NoteModel
-from brain_brew.representation.yaml.part_holder import PartHolder
+from brain_brew.representation.yaml.note_model import NoteModel
 
 
 @dataclass
