+++ brain_brew/build_tasks/crowd_anki/note_models_to_crowd_anki.py
@@ -1,10 +1,10 @@
 from dataclasses import dataclass, field
 from typing import Union, List
 
+from brain_brew.configuration.part_holder import PartHolder
+from brain_brew.configuration.representation_base import RepresentationBase
 from brain_brew.interfaces.yamale_verifyable import YamlRepr
-from brain_brew.representation.configuration.representation_base import RepresentationBase
-from brain_brew.representation.yaml.note_model_repr import NoteModel
-from brain_brew.representation.yaml.part_holder import PartHolder
+from brain_brew.representation.yaml.note_model import NoteModel
 
 
 @dataclass
