+++ brain_brew/build_tasks/crowd_anki/notes_from_crowd_anki.py
@@ -3,12 +3,12 @@ from dataclasses import dataclass, field
 from typing import Union, Optional, List
 
 from brain_brew.build_tasks.crowd_anki.shared_base_notes import SharedBaseNotes
-from brain_brew.representation.build_config.build_task import BuildPartTask
-from brain_brew.representation.configuration.representation_base import RepresentationBase
+from brain_brew.configuration.build_config.build_task import BuildPartTask
+from brain_brew.configuration.part_holder import PartHolder
+from brain_brew.configuration.representation_base import RepresentationBase
 from brain_brew.representation.json.crowd_anki_export import CrowdAnkiExport
 from brain_brew.representation.json.wrappers_for_crowd_anki import CrowdAnkiJsonWrapper, CrowdAnkiNoteWrapper
-from brain_brew.representation.yaml.note_repr import Notes, Note
-from brain_brew.representation.yaml.part_holder import PartHolder
+from brain_brew.representation.yaml.notes import Notes, Note
 
 
 @dataclass
