+++ brain_brew/build_tasks/crowd_anki/notes_to_crowd_anki.py
@@ -2,11 +2,11 @@ from dataclasses import dataclass, field
 from typing import Optional, Union, List
 
 from brain_brew.build_tasks.crowd_anki.shared_base_notes import SharedBaseNotes
+from brain_brew.configuration.part_holder import PartHolder
+from brain_brew.configuration.representation_base import RepresentationBase
 from brain_brew.interfaces.yamale_verifyable import YamlRepr
-from brain_brew.representation.configuration.representation_base import RepresentationBase
 from brain_brew.representation.json.wrappers_for_crowd_anki import CrowdAnkiNoteWrapper
-from brain_brew.representation.yaml.note_repr import Notes, Note
-from brain_brew.representation.yaml.part_holder import PartHolder
+from brain_brew.representation.yaml.notes import Notes, Note
 from brain_brew.utils import blank_str_if_none
