+++ brain_brew/build_tasks/csvs/csvs_generate.py
@@ -2,11 +2,11 @@ from dataclasses import dataclass, field
 from typing import List, Dict, Union
 
 from brain_brew.build_tasks.csvs.shared_base_csvs import SharedBaseCsvs
-from brain_brew.representation.build_config.build_task import TopLevelBuildTask
-from brain_brew.representation.configuration.csv_file_mapping import FileMapping
-from brain_brew.representation.configuration.note_model_mapping import NoteModelMapping
-from brain_brew.representation.yaml.note_repr import Notes, Note
-from brain_brew.representation.yaml.part_holder import PartHolder
+from brain_brew.configuration.build_config.build_task import TopLevelBuildTask
+from brain_brew.configuration.part_holder import PartHolder
+from brain_brew.representation.yaml.notes import Notes, Note
+from brain_brew.transformers.file_mapping import FileMapping
+from brain_brew.transformers.note_model_mapping import NoteModelMapping
 from brain_brew.utils import join_tags
