+++ brain_brew/build_tasks/deck_parts/from_yaml_part.py
@@ -2,13 +2,13 @@ from abc import ABCMeta
 from dataclasses import dataclass
 from typing import Union
 
-from brain_brew.representation.build_config.build_task import BuildPartTask
-from brain_brew.representation.configuration.representation_base import RepresentationBase
-from brain_brew.representation.yaml.headers_repr import Headers
-from brain_brew.representation.yaml.media_group_repr import MediaGroup
-from brain_brew.representation.yaml.note_model_repr import NoteModel
-from brain_brew.representation.yaml.note_repr import Notes
-from brain_brew.representation.yaml.part_holder import PartHolder
+from brain_brew.configuration.build_config.build_task import BuildPartTask
+from brain_brew.configuration.part_holder import PartHolder
+from brain_brew.configuration.representation_base import RepresentationBase
+from brain_brew.representation.yaml.headers import Headers
+from brain_brew.representation.yaml.media_group import MediaGroup
+from brain_brew.representation.yaml.note_model import NoteModel
+from brain_brew.representation.yaml.notes import Notes
 from brain_brew.representation.yaml.yaml_object import YamlObject
