+++ brain_brew/build_tasks/deck_parts/media_group_to_folder.py
@@ -1,10 +1,10 @@
 from dataclasses import dataclass
 from typing import List, Union, Optional
 
-from brain_brew.representation.build_config.build_task import BuildPartTask
-from brain_brew.representation.configuration.representation_base import RepresentationBase
-from brain_brew.representation.yaml.media_group_repr import MediaGroup
-from brain_brew.representation.yaml.part_holder import PartHolder
+from brain_brew.configuration.build_config.build_task import BuildPartTask
+from brain_brew.configuration.part_holder import PartHolder
+from brain_brew.configuration.representation_base import RepresentationBase
+from brain_brew.representation.yaml.media_group import MediaGroup
 from brain_brew.transformers.media_group_save_to_location import save_media_groups_to_location
