+++ brain_brew/build_tasks/deck_parts/note_model_template_from_html_files.py
@@ -1,11 +1,11 @@
 from dataclasses import dataclass, field
 from typing import Optional
 
-from brain_brew.representation.build_config.build_task import BuildPartTask
-from brain_brew.representation.configuration.representation_base import RepresentationBase
+from brain_brew.configuration.build_config.build_task import BuildPartTask
+from brain_brew.configuration.part_holder import PartHolder
+from brain_brew.configuration.representation_base import RepresentationBase
 from brain_brew.representation.generic.html_file import HTMLFile
 from brain_brew.representation.yaml.note_model_template import Template
-from brain_brew.representation.yaml.part_holder import PartHolder
 
 html_separator = '--'
