+++ brain_brew/configuration/build_config/parts_builder.py
@@ -1,23 +1,19 @@
-from abc import ABCMeta
 from dataclasses import dataclass
-from typing import Dict, Type, List, Tuple, Set
-from textwrap import dedent
+from typing import Dict, Type, List, Set
 
-# Build Tasks
-from brain_brew.build_tasks.csvs.notes_from_csvs import NotesFromCsvs
-from brain_brew.build_tasks.deck_parts.from_yaml_part import NotesFromYamlPart, HeadersFromYamlPart, \
-    NoteModelsFromYamlPart, MediaGroupFromYamlPart
-from brain_brew.build_tasks.deck_parts.media_group_from_folder import MediaGroupFromFolder
 from brain_brew.build_tasks.crowd_anki.headers_from_crowdanki import HeadersFromCrowdAnki
 from brain_brew.build_tasks.crowd_anki.media_group_from_crowd_anki import MediaGroupFromCrowdAnki
 from brain_brew.build_tasks.crowd_anki.note_models_from_crowd_anki import NoteModelsFromCrowdAnki
 from brain_brew.build_tasks.crowd_anki.notes_from_crowd_anki import NotesFromCrowdAnki
+from brain_brew.build_tasks.csvs.notes_from_csvs import NotesFromCsvs
+from brain_brew.build_tasks.deck_parts.from_yaml_part import NotesFromYamlPart, HeadersFromYamlPart, \
+    NoteModelsFromYamlPart, MediaGroupFromYamlPart
+from brain_brew.build_tasks.deck_parts.media_group_from_folder import MediaGroupFromFolder
 from brain_brew.build_tasks.deck_parts.media_group_to_folder import MediaGroupsToFolder
 from brain_brew.build_tasks.deck_parts.note_model_from_html_parts import NoteModelFromHTMLParts
 from brain_brew.build_tasks.deck_parts.note_model_template_from_html_files import TemplateFromHTML
-
-from brain_brew.representation.build_config.build_task import BuildTask, BuildPartTask, TopLevelBuildTask
-from brain_brew.representation.build_config.recipe_builder import RecipeBuilder
+from brain_brew.configuration.build_config.build_task import BuildTask, BuildPartTask, TopLevelBuildTask
+from brain_brew.configuration.build_config.recipe_builder import RecipeBuilder
 
 
 @dataclass
