+++ brain_brew/configuration/build_config/recipe_builder.py
@@ -1,10 +1,10 @@
 import re
 from abc import ABCMeta, abstractmethod
 from dataclasses import dataclass
+from textwrap import indent
 from typing import Dict, List, Type, Set
-from textwrap import indent, dedent
 
-from brain_brew.representation.build_config.build_task import BuildTask
+from brain_brew.configuration.build_config.build_task import BuildTask
 from brain_brew.representation.yaml.yaml_object import YamlObject
