+++ brain_brew/configuration/global_config.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass, field
 from typing import Union, Optional
 
-from brain_brew.representation.configuration.representation_base import RepresentationBase
+from brain_brew.configuration.representation_base import RepresentationBase
 from brain_brew.representation.yaml.yaml_object import YamlObject
