+++ brain_brew/configuration/representation_base.py
