+++ brain_brew/representation/json/crowd_anki_export.py
@@ -5,7 +5,7 @@ from typing import List
 from brain_brew.representation.generic.source_file import SourceFile
 from brain_brew.representation.json.json_file import JsonFile
 from brain_brew.representation.json.wrappers_for_crowd_anki import CrowdAnkiJsonWrapper
-from brain_brew.representation.yaml.note_model_repr import NoteModel
+from brain_brew.representation.yaml.note_model import NoteModel
 from brain_brew.utils import create_path_if_not_exists
