+++ brain_brew/transformers/note_model_mapping.py
@@ -2,11 +2,11 @@ from dataclasses import dataclass
 from enum import Enum
 from typing import List, Union, Dict
 
+from brain_brew.configuration.part_holder import PartHolder
+from brain_brew.configuration.representation_base import RepresentationBase
 from brain_brew.interfaces.yamale_verifyable import YamlRepr
-from brain_brew.representation.configuration.representation_base import RepresentationBase
-from brain_brew.representation.yaml.note_model_repr import NoteModel
-from brain_brew.representation.yaml.note_repr import GUID, TAGS
-from brain_brew.representation.yaml.part_holder import PartHolder
+from brain_brew.representation.yaml.note_model import NoteModel
+from brain_brew.representation.yaml.notes import GUID, TAGS
 from brain_brew.utils import single_item_to_list
