+++ scripts/yamale_build.py
@@ -1,8 +1,9 @@
-import sys
 import os
+import sys
+
 sys.path.append(os.path.abspath(''))
 
-from brain_brew.representation.build_config.top_level_builder import TopLevelBuilder
+from brain_brew.configuration.build_config.top_level_builder import TopLevelBuilder
 
 build: str = TopLevelBuilder.build_yamale()
 filepath = "brain_brew/schemas/recipe.yaml"
